♻️ refactor(types): move agentCronJob schemas to types package
- Create manual Zod schemas in @lobechat/types instead of using createInsertSchema
- Define InsertAgentCronJobSchema and UpdateAgentCronJobSchema manually
- Re-export types from database schema for backward compatibility
- Update router to use new schema imports from @lobechat/types
