+++ packages/types/src/index.ts
@@ -1,4 +1,5 @@
 export * from './agent';
+export * from './agentCronJob';
 export * from './agentGroup';
 export * from './aiChat';
 export * from './aiProvider';
